Исправлена ошибка валидации
