+++ js/create-photo.js
@@ -6,21 +6,13 @@ const similar_template = document.querySelector('#picture').content.querySelecto
 const fragment = document.createDocumentFragment();
 
 
-
 arrayOfObjectsPhoto.forEach((item) => {
   const picture_template = similar_template.cloneNode(true);
   picture_template.querySelector('.picture__img').src = item.url;
   picture_template.querySelector('.picture__likes').textContent = item.likes;
   picture_template.querySelector('.picture__comments').textContent = item.message;
   picture_template.id = item.id;
-
   fragment.appendChild(picture_template);
 });
 
-
-
-
 picture_container.appendChild(fragment);
-
-
-window.console.log(arrayOfObjectsPhoto);
