+++ js/valid-hashtags.js
@@ -1,12 +1,6 @@
-import {isEscEvent} from './utils.js';
-
-
 const MAX_HASHTAGS = 5;
-const HASHTAG_VALIDITY_REGEX = RegExp(`^#[a-zA-Z0-9а-яА-ЯёЁ]{1,19}$`);
-
+const HASHTAG_VALIDITY_REGEX = RegExp('^#[a-zA-Z0-9а-яА-ЯёЁ]{1,19}$');
 const fieldHashtagElement = document.querySelector('.text__hashtags');
-
-
 
 
 const validateHashtags = function() {
@@ -14,34 +8,22 @@ const validateHashtags = function() {
   fieldHashtagElement.addEventListener('input', function(evt) {
 
     const hashtagsArray = evt.target.value.toLowerCase().split(' ');
-
     const isDuplicateHashtagInArray = !hashtagsArray.every((item, index, array) => array.indexOf(item) === index);
-
-    const isInvalidHashtagInArray = !hashtagsArray.every((item) => (item === ' ' || HASHTAG_VALIDITY_REGEX.test(item)));
+    const isInvalidHashtagInArray = !hashtagsArray.every((item) => (item === '' || HASHTAG_VALIDITY_REGEX.test(item)));
 
 
     if(hashtagsArray.length > MAX_HASHTAGS){
       fieldHashtagElement.setCustomValidity('Кол-во хэштегов не может превышать 5-ти')
     } else if(isInvalidHashtagInArray){
-      fieldHashtagElement.setCustomValidity('Хэштеги не могут содержать цифры,знаки,эмодзи и начинатся с #')
+      fieldHashtagElement.setCustomValidity('Хэштеги не могут содержать цифры, знаки, эмодзи и начинаться с #')
     } else if(isDuplicateHashtagInArray){
       fieldHashtagElement.setCustomValidity('Хэштеги недолжны повторяться')
-    } else{
+    } else {
       fieldHashtagElement.setCustomValidity('')
     }
+
+    fieldHashtagElement.reportValidity();
   });
 }
 
-
-const onFieldFocus = function(evt) {
-
-  evt.target.addEventListener('keydown', function(evt) {
-
-    if(isEscEvent(evt)) {
-      evt.preventDefault();
-      evt.stopPropagation();
-    }
-  });
-};
-
-export {validateHashtags, onFieldFocus};
+export {validateHashtags};
